src - Copie/test/javascript/spec/app/entities/lignes-ventes/lignes-ventes.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { LignesVentesService } from 'app/entities/lignes-ventes/lignes-ventes.service';
import { ILignesVentes, LignesVentes } from 'app/shared/model/lignes-ventes.model';

describe('Service Tests', () => {
  describe('LignesVentes Service', () => {
    let injector: TestBed;
    let service: LignesVentesService;
    let httpMock: HttpTestingController;
    let elemDefault: ILignesVentes;
    let expectedResult: ILignesVentes | ILignesVentes[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(LignesVentesService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new LignesVentes(0, 0, 0, 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a LignesVentes', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new LignesVentes()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a LignesVentes', () => {
        const returnedFromService = Object.assign(
          {
            prixUnitaire: 1,
            quantite: 1,
            prixTotal: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of LignesVentes', () => {
        const returnedFromService = Object.assign(
          {
            prixUnitaire: 1,
            quantite: 1,
            prixTotal: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a LignesVentes', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
